update(HVRoCNpc + HVRoCItem): Extracted special logic for vob types to separate components.
