+++ Assets/Gothic-UnZENity-Core/Resources/Prefabs/Vobs/oCNpc.prefab.meta
