+++ Assets/Gothic-UnZENity-HVR/Resources/HVR/Prefabs/Vobs/oCNpc.prefab.meta
