+++ Assets/Gothic-UnZENity-HVR/Scripts/Components/HVRoCNpc.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9df2797e60b740a184fb2cfd36d0a490
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
